+++ packages/eas-cli/src/eas-update/utils.ts
@@ -1,0 +1,15 @@
+import { ExpoConfig } from '@expo/config-types';
+
+import { ExpoGraphqlClient } from '../commandUtils/context/contextUtils/createGraphqlClient';
+
+export type EASUpdateContext = {
+  graphqlClient: ExpoGraphqlClient;
+  nonInteractive: boolean;
+  app: { exp: ExpoConfig; projectId: string };
+};
+
+export interface EASUpdateAction<T = any> {
+  runAsync(ctx: EASUpdateContext): Promise<T>;
+}
+
+export class NonInteractiveError extends Error {}